==> 2_features_creation/2_Magic Features (0.03 gain)  Commented.py <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load in 
laptop = 0
if laptop == 1:
    drive = "C"
else:
    drive = "F"

import numpy as np
import pandas as pd
import timeit



if laptop == 1:
    train_orig =  pd.read_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/train.csv', header=0)
    test_orig =  pd.read_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/test.csv', header=0)
else:
    train_orig =  pd.read_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/train.csv', header=0)
    test_orig =  pd.read_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/test.csv', header=0)

#copy only one column of question for train and test
tic0=timeit.default_timer()
df1 = train_orig[['question1']].copy()
df2 = train_orig[['question2']].copy()
df1_test = test_orig[['question1']].copy()
df2_test = test_orig[['question2']].copy()

df2.rename(columns = {'question2':'question1'},inplace=True)
df2_test.rename(columns = {'question2':'question1'},inplace=True)

#concat all the questions and remove duplicates
train_questions = df1.append(df2)
train_questions = train_questions.append(df1_test)
train_questions = train_questions.append(df2_test)
#train_questions.drop_duplicates(subset = ['qid1'],inplace=True)
train_questions.drop_duplicates(subset = ['question1'],inplace=True)

#create a big dict of questions to index
train_questions.reset_index(inplace=True,drop=True)
questions_dict = pd.Series(train_questions.index.values,index=train_questions.question1.values).to_dict()

#copy all original, remove useless columns
train_cp = train_orig.copy()
test_cp = test_orig.copy()
train_cp.drop(['qid1','qid2'],axis=1,inplace=True)

#formating testcp to be like traincp and concat
test_cp['is_duplicate'] = -1
test_cp.rename(columns={'test_id':'id'},inplace=True)
comb = pd.concat([train_cp,test_cp])

#get id from the dict
comb['q1_hash'] = comb['question1'].map(questions_dict)
comb['q2_hash'] = comb['question2'].map(questions_dict)


q1_vc = comb.q1_hash.value_counts().to_dict()
q2_vc = comb.q2_hash.value_counts().to_dict()

def try_apply_dict(x,dict_to_apply):
    try:
        return dict_to_apply[x]
    except KeyError:
        return 0
#map to frequency space
comb['q1_freq'] = comb['q1_hash'].map(lambda x: try_apply_dict(x,q1_vc) + try_apply_dict(x,q2_vc))
comb['q2_freq'] = comb['q2_hash'].map(lambda x: try_apply_dict(x,q1_vc) + try_apply_dict(x,q2_vc))

train_comb = comb[comb['is_duplicate'] >= 0][['id','q1_hash','q2_hash','q1_freq','q2_freq','is_duplicate']]
test_comb = comb[comb['is_duplicate'] < 0][['id','q1_hash','q2_hash','q1_freq','q2_freq']]

comb.head()

corr_mat = train_comb.corr()
corr_mat.head()
#more frequenct questions are more likely to be duplicates

train_comb[['id', 'q1_freq', 'q2_freq']].to_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/train_comb.csv') 
test_comb[['id', 'q1_freq', 'q2_freq']].to_csv(drive + ':/DS-main/Kaggle-main/Quora Question Pairs - inputs/data/test_comb.csv')
